replaced the name of the files and folders, placed all DOM elements && global variables into their own modules, PLUS I added some setter functions for those global vars. Also updated the readme for the new link
