+++ .gitignore
@@ -1,2 +1,3 @@
-.env
-.txt
+*.env
+*.txt
+node_modules
