+++ css/partials/gallery/_main.scss
@@ -13,7 +13,6 @@ main{
   background-color: $clr-secondary-bg;
   margin: 1.2rem auto 1.25rem;
   width: 90%;
-  min-height: 100vh;
 }
 
 .banner{
@@ -22,5 +21,12 @@ main{
   font-size: 1.5em;
   color: $clr-secondary-txt;
   background-color: wheat;
+}
+
+.unshoved{
+  min-height: calc(100vh - 51px - 66px - 2.45rem);
+}
+.shoved{
+  min-height: calc(100vh - 94px - 66px - 2.45rem);
 }
 //footer styles
