+++ css/partials/global/_header.scss
@@ -29,7 +29,6 @@ header{
 }
 
 
-
 header[data-gallery-header] {
   z-index: 3;
 
@@ -44,6 +43,7 @@ header[data-gallery-header] {
   justify-content: space-around;
   flex-flow: row wrap;
   margin: 0 auto;
+
   h1 {
     display: none;
     font-family: $secondary-font;
@@ -56,7 +56,7 @@ header[data-gallery-header] {
     width: 50%;
     align-items: center;
     display: flex;
-
+  
     form {
       width: 100%;
       height: 100%;
@@ -64,33 +64,37 @@ header[data-gallery-header] {
       display: flex;
       justify-content: flex-start;
       border-radius: 50px 50px;
-      button {
-        font-size: inherit;
-        border: none;
-        width: fit-content;
-        height: 100%;
-        padding: 10px;
-        color: $clr-secondary-txt;
-        background: $clr-secondary-bg;
-        &:hover{
-          cursor: pointer;
-          color: $clr-primary-txt;
-          background-color: $clr-primary-bg-lighter;
-        }
+      
+    }
+  
+    button {
+      font-size: inherit;
+      border: none;
+      width: fit-content;
+      height: 100%;
+      padding: 10px;
+      color: $clr-secondary-txt;
+      background: $clr-secondary-bg;
+      &:hover{
+        cursor: pointer;
+        color: $clr-primary-txt;
+        background-color: $clr-primary-bg-lighter;
       }
-      input {
-        font-size: inherit;
-        width: 100%;
-        height: 100%;
-        border: none;
-        padding:0 5px;
-        color: $clr-secondary-txt;
-      }
+    }
+  
+    input {
+      font-size: inherit;
+      width: 100%;
+      height: 100%;
+      border: none;
+      padding:0 5px;
+      color: $clr-secondary-txt;
     }
   }
 
-
 }
+
+
 
 .bar {
   width: 35px;
@@ -135,4 +139,5 @@ nav{
 .autoHeight{
   display: block;
   height: auto;
+  padding-top: 10px;
 }
