+++ gallery.html
@@ -14,6 +14,7 @@
     <link rel="preload" as="font" type="font/woff2"  href="webfonts/fa-brands-400.woff2" crossorigin/>
     <link rel="preload" as="font" type="font/woff2"  href="webfonts/fa-solid-900.woff2" crossorigin/>
 
+    <link rel="icon" href ="/favicon.ico" type="image/x-icon" /> <link rel="shortcut icon" href="/favicon.ico" type="image/x-icon" />
     
   </head>
 
@@ -79,12 +80,9 @@
 
 
     <main id="index-main">
-      <div id="grid-container">
-        <div class="title">
-          <h2 id="anime-title"></h2>
-        </div>
+      <div id="grid-container" class="unshoved">
 
-        <div class="open-grid" id="anime-page"></div>
+        <div class="open-grid" id="Images-container"></div>
 
       </div>
       
@@ -100,7 +98,7 @@
       </ul>
       
     </footer>
-    <script type="module" src="script/prototype.js"></script>
+    <script type="module" src="js/app.js"></script>
     <link rel="shortcut icon" href="#">
