--- script/ConstantData.js
@@ -1,1 +1,0 @@
-export let initialSearch = new URLSearchParams(window.location.search).get("search")
