--- script/CounterVars.js
@@ -1,10 +1,0 @@
-//export let totalRendered = 0;
-
-// export function increment(counter, value){
-//     counter += value
-// }
-
-// const CounterVars = {
-//     totalRendered: 0,
-// }
-// export default CounterVars;
