--- script/Elements.js
@@ -1,41 +1,0 @@
-export const category = document.getElementById("anime-page");
-
-export const category_Title = document.getElementById("anime-title");
-
-export const body = document.getElementById("body");
-
-export const modalBox = document.getElementById("modal-box");
-
-export const ModalImage = document.getElementById("modal-img");
-
-export const photographer = document.getElementById('photographer');
-export const photoLink = document.getElementById('photo-link');
-
-export const form = document.getElementsByTagName('form');
-export const input = document.getElementById("search-input");
-
-export const banner = document.querySelector('.banner');
-
-
-// const ConstVars = {
-//   category : document.getElementById("anime-page"),
-
-//   category_Title : document.getElementById("anime-title"),
-
-//   body : document.getElementById("body"),
-
-//   modalBox : document.getElementById("modal-box"),
-
-//   ModalImage : document.getElementById("modal-img"),
-
-//   photographer : document.getElementById('photographer'),
-//   photoLink : document.getElementById('photo-link'),
-
-//   form : document.getElementsByTagName('form'),
-//   input : document.getElementById("search-input"),
-
-//   banner : document.querySelector('.banner'),
-
-// }
-
-// export default ConstVars;
