--- script/steps.txt
@@ -1,12 +1,0 @@
-expected
-
-onScroll - render new images, add functionality to the images.
-
-
-
-
-
-updataValue('data1')
-return data1.next_page
-let currentData = updataValue('data1')
-updataValue('currentData')
